Breakthrough on Returns from Nested Functions
Lost in the midst of picking up Javascript, WebDriver, and Protractor all at once: Returning values from nested functions. In the LogOut function in home-page.js, I was not properly returning the elementFinder when searching for the button. It makes sense now that I'm looking at the correct way, but it LOOKED close enough with the way I had it before that I didn't realize what I was doing wrong.